refactor: economy phase-2 slice 11 — build-cost scenario tables (#3856)
Migrate build_cost_test.dart to table-driven scenarios in test_support,
preserving all nine test descriptions for duplicate-description CI.
